Add Makefile_dev for compiling/loading
The production Makefile cleans (i.e. removes) all  files in all subdirectories in the targets/*/armgcc.
This includes the appropriate .hex, .zip and .map files and the _build sub directory containing nrf*_xxaa.in, .out,. map,.hex., bin.
Unfortunately it unnecessarily removes all compiled object files in the targets/*/armgcc/_build/nrf*_xxaa/ subdirectory.
This causes the compilation step to recompile all sources, even those which are unchanged (most of the advantage of using make in the first place).
In addition all variations of boards including kaarle, kalervo and keijo are relinked. 

This makefile_dev only removes object files that need to be recompiled because of date stamping. and only links ruuvitab_b.
Not only is this much faster but less noisy and problems are less likely to be hidden .

It maybe (?) necessary to rm the  targets/*/armgcc/_build/nrf*_xxaa/ subdirectory. manually if a macro library or other source file that is not noticed by make is changed.